[DEPYLER-0275] Fix Code Quality: Remove Unnecessary CSE Temps, Lifetimes, and Casts
Applied Extreme TDD methodology to fix 3 code quality issues discovered during
matrix testing validation of python-to-rust-conversion-examples.

**CSE Temp Elision** (50% reduction)
- Before: `let _cse_temp_0 = x + y; _cse_temp_0`
- After: `x + y` (direct return)
- Root cause: Optimizer applied CSE to final returns without checking simplicity
- Fix: Added `is_simple_return_expr()` check in optimizer.rs:530-714
- Impact: Reduced CSE temps from 21→10 in basic_types example

**Lifetime Elision** (cleaner signatures)
- Before: `pub fn f<'a>(s: &'a str) -> i32`
- After: `pub fn f(s: &str) -> i32`
- Root cause: Always generated explicit lifetimes even when Rust could elide
- Fix: Integrated `apply_elision_rules()` in func_gen.rs:626-630
- Fix: Added `should_elide_lifetimes` parameter to borrowing logic

**Double Cast Removal**
- Before: `s.len() as i32 as i32`
- After: `s.len() as i32`
- Root cause: `convert_len_call()` added cast, then caller added another
- Fix: Removed premature cast from expr_gen.rs:690-701

**Testing Approach**
- RED: Created minimal failing test cases demonstrating each issue
- GREEN: Fixed root cause in transpiler, not generated output
- REFACTOR: Verified all changes with cargo clippy -- -D warnings

**Result**
- Generated Rust code now passes clippy with zero warnings
- Cleaner, more idiomatic output
- Follows Rust's elision rules and best practices

Files modified:
- crates/depyler-core/src/optimizer.rs (CSE elision)
- crates/depyler-core/src/lifetime_analysis.rs (elision rules)
- crates/depyler-core/src/rust_gen/func_gen.rs (lifetime application)
- crates/depyler-core/src/rust_gen/expr_gen.rs (cast removal)
- CHANGELOG.md (documented changes)
- docs/bugs/DEPYLER-0275-ROOT-CAUSE-ANALYSIS.md (detailed analysis)

+++ crates/depyler-core/src/rust_gen/expr_gen.rs
@@ -693,14 +693,11 @@ impl<'a, 'b> ExpressionConverter<'a, 'b> {
         }
         let arg = &args[0];
 
-        // Python's len() returns int, which we map to i32/i64/isize based on type mapper.
-        // Rust's .len() returns usize, so we cast to match Python's int type.
-        // This ensures type consistency: len() - 1, len() comparisons, etc. all work with i32.
-        //
-        // Note: This matches the type mapper's integer width preference.
-        // For functions returning indices, they should explicitly use usize in their return type.
-        // Removed outer parens - they're unnecessary and cause clippy warnings
-        Ok(parse_quote! { #arg.len() as i32 })
+        // DEPYLER-0275: Return .len() without cast - let caller add cast if needed
+        // Python's len() → Rust's .len() which returns usize
+        // The return statement processing in stmt_gen.rs will add `as i32` if needed
+        // This prevents double casts like `s.len() as i32 as i32`
+        Ok(parse_quote! { #arg.len() })
     }
 
     fn convert_int_cast(&self, hir_args: &[HirExpr], arg_exprs: &[syn::Expr]) -> Result<syn::Expr> {
